fix: перенести запросы регистрации и логина в отдельный файл auth.js
