--- src/pages/RegisterPage/AuthApi.js
@@ -1,21 +1,0 @@
-import { baseUrl } from '../../utils/constants';
-
-const checkServerResponse = (res) => {
-  return res.ok
-    ? res.json()
-    // eslint-disable-next-line prefer-promise-reject-errors
-    : Promise.reject(`Ошибка: ${res.status}`);
-};
-
-// eslint-disable-next-line import/prefer-default-export
-export const register = (username, password, email) => {
-  return fetch(`${baseUrl}/auth/users/`, {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ username, password, email }),
-  })
-    .then(checkServerResponse);
-};
